feat: add video embed hook and utility
