+++ components/content/content-detail-media.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { useState } from 'react'
 import type { ContentItem } from '@/lib/types/database'
+import { useVideoEmbed } from '@/lib/hooks/useVideoEmbed'
 
 interface ContentDetailMediaProps {
   content: ContentItem
@@ -15,21 +15,8 @@ interface ContentDetailMediaProps {
  * - Audio player for audio content
  */
 export function ContentDetailMedia({ content }: ContentDetailMediaProps) {
-  const [videoLoaded, setVideoLoaded] = useState(false)
   const videoUrl = content.metadata?.mediaUrl || content.metadata?.embed_links?.[0]
-
-  const getEmbedUrl = (url: string) => {
-    if (!url) return ''
-    const ytMatch = url.match(/(?:youtu\.be\/|youtube\.com\/(?:watch\?v=|embed\/|v\/|shorts\/))([\w-]{11})/i)
-    if (ytMatch) {
-      return `https://www.youtube.com/embed/${ytMatch[1]}`
-    }
-    const vimeoMatch = url.match(/vimeo\.com\/(?:video\/)?(\d+)/)
-    if (vimeoMatch) {
-      return `https://player.vimeo.com/video/${vimeoMatch[1]}`
-    }
-    return url
-  }
+  const { embedUrl, videoLoaded, handleLoad } = useVideoEmbed(videoUrl)
 
   return (
     <div className="bg-white rounded-lg shadow-md p-6 mb-8">
@@ -41,11 +28,11 @@ export function ContentDetailMedia({ content }: ContentDetailMediaProps) {
             </div>
           )}
           <iframe
-            src={getEmbedUrl(videoUrl)}
+            src={embedUrl}
             className={`w-full h-full rounded-lg ${videoLoaded ? '' : 'hidden'}`}
             allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
             allowFullScreen
-            onLoad={() => setVideoLoaded(true)}
+            onLoad={handleLoad}
           />
         </div>
       )}
